Obtaining user id

+++ src/main/kotlin/com/artemissoftware/calliopenotes/controllers/note/NoteController.kt
@@ -6,6 +6,7 @@ import com.artemissoftware.calliopenotes.controllers.note.models.NoteRequest
 import com.artemissoftware.calliopenotes.controllers.note.models.NoteResponse
 import com.artemissoftware.calliopenotes.database.repository.NoteRepository
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -18,14 +19,14 @@ class NoteController(
     fun save(
         /*@Valid*/ @RequestBody body: NoteRequest
     ): NoteResponse {
-        val ownerId = "1"//SecurityContextHolder.getContext().authentication.principal as String
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = repository.save(body.toNote(ownerId))
         return note.toResponse()
     }
 
     @GetMapping
     fun findByOwnerId(): List<NoteResponse> {
-        val ownerId = "1"//SecurityContextHolder.getContext().authentication.principal as String
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return repository.findByOwnerId(ObjectId(ownerId)).map {
             it.toResponse()
         }
@@ -36,9 +37,9 @@ class NoteController(
         val note = repository.findById(ObjectId(id)).orElseThrow {
             IllegalArgumentException("Note not found")
         }
-        //val ownerId = SecurityContextHolder.getContext().authentication.principal as String
-        //if(note.ownerId.toHexString() == ownerId) {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        if(note.ownerId.toHexString() == ownerId) {
             repository.deleteById(ObjectId(id))
-        //}
+        }
     }
 }
